+++ hl7v2-readyapi-plugin/src/main/java/ca/bc/gov/hlth/hl7v2plugin/teststeps/panels/EncodeHL7v2TestStepPanel.java
@@ -2,8 +2,10 @@ package ca.bc.gov.hlth.hl7v2plugin.teststeps.panels;
 
 import java.awt.BorderLayout;
 import java.awt.Font;
+import java.awt.event.KeyEvent;
 import java.util.Date;
 
+import javax.swing.JButton;
 import javax.swing.JComponent;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
@@ -15,9 +17,11 @@ import org.slf4j.LoggerFactory;
 
 import com.eviware.soapui.impl.wsdl.panels.teststeps.common.TestStepVariables;
 import com.eviware.soapui.support.DateUtil;
+import com.eviware.soapui.support.UISupport;
 import com.eviware.soapui.support.components.JComponentInspector;
 import com.eviware.soapui.support.components.JInspectorPanel;
 import com.eviware.soapui.support.components.JInspectorPanelFactory;
+import com.eviware.soapui.support.components.JXToolBar;
 import com.eviware.soapui.support.components.SimpleBindingForm;
 import com.eviware.soapui.support.log.JLogList;
 import com.eviware.soapui.ui.support.ModelItemDesktopPanel;
@@ -29,6 +33,7 @@ import ca.bc.gov.hlth.hl7v2plugin.teststeps.ExecutableTestStep;
 import ca.bc.gov.hlth.hl7v2plugin.teststeps.ExecutableTestStepResult;
 import ca.bc.gov.hlth.hl7v2plugin.teststeps.ExecutionListener;
 import ca.bc.gov.hlth.hl7v2plugin.teststeps.PublishTestStep;
+import ca.bc.gov.hlth.hl7v2plugin.teststeps.actions.RunTestStepAction;
 
 public class EncodeHL7v2TestStepPanel extends ModelItemDesktopPanel<EncodeHL7v2TestStep> implements ExecutionListener {
     
@@ -39,6 +44,7 @@ public class EncodeHL7v2TestStepPanel extends ModelItemDesktopPanel<EncodeHL7v2T
 	private JInspectorPanel inspectorPanel;
     private JComponentInspector<JComponent> logInspector;
     private JLogList logArea;
+    private RunTestStepAction startAction;
 
     public EncodeHL7v2TestStepPanel(EncodeHL7v2TestStep modelItem) {
         super(modelItem);
@@ -87,11 +93,25 @@ public class EncodeHL7v2TestStepPanel extends ModelItemDesktopPanel<EncodeHL7v2T
 
         JLabel instructionsLabel = new JLabel("This step encodes the Data Source fileContent as HL7v2 and makes it available as the testcase property ${#TestCase#encodedHL7v2Msg}");
         form.append(instructionsLabel);
-
+        
         JPanel result = new JPanel(new BorderLayout(0, 0));
         result.add(new JScrollPane(form.getPanel(), ScrollPaneConstants.VERTICAL_SCROLLBAR_AS_NEEDED,
                 ScrollPaneConstants.HORIZONTAL_SCROLLBAR_AS_NEEDED), BorderLayout.CENTER);
+
+        result.add(buildToolbar(), BorderLayout.NORTH);
         return result;
+    }
+    
+    private JComponent buildToolbar() {
+        JXToolBar toolBar = UISupport.createToolbar();
+        startAction = new RunTestStepAction(getModelItem());
+        JButton submitButton = UISupport.createActionButton(startAction, startAction.isEnabled());
+        toolBar.add(submitButton);
+        submitButton.setMnemonic(KeyEvent.VK_ENTER);
+        toolBar.add(UISupport.createActionButton(startAction.getCorrespondingStopAction(), startAction
+                .getCorrespondingStopAction().isEnabled()));
+        //addConnectionActionsToToolbar(toolBar);
+        return toolBar;
     }
 
     @Override